+++ src/main/java/com/vizaco/onlinecontrol/dao/SheduleDao.java
@@ -4,27 +4,26 @@ import com.vizaco.onlinecontrol.model.*;
 import com.vizaco.onlinecontrol.representation.JournalView;
 import org.springframework.dao.DataAccessException;
 
-import java.time.DayOfWeek;
 import java.util.Date;
 import java.util.List;
 
 public interface SheduleDao {
 
-    Shedule findSheduleById(Long id) throws DataAccessException;
+    Shedule findSheduleById(Integer id) throws DataAccessException;
 
     List<Shedule> getSheduleBeetwenAnyCriteria(Date start, Date end, Clazz clazz, Period period, Subject subject, Teacher teacher) throws DataAccessException;
 
     List<JournalView> getJournalByCriteria(Date start, Date end) throws DataAccessException;
 
-    Period findPeriodById(Long id) throws DataAccessException;
+    Period findPeriodById(Integer id) throws DataAccessException;
 
     List<Period> getAllPeriods() throws DataAccessException;
 
-    Subject findSubjectById(Long id) throws DataAccessException;
+    Subject findSubjectById(Integer id) throws DataAccessException;
 
     List<Subject> getAllSubjects() throws DataAccessException;
 
-    Teacher findTeacherById(Long id) throws DataAccessException;
+    Teacher findTeacherById(Integer id) throws DataAccessException;
 
     List<Teacher> getAllTeachers() throws DataAccessException;
